+++ apps/io-func-sign-issuer/src/infra/azure/event-hubs/event.ts
